+++ src/components/StackedAreaChart/components/index.ts
@@ -1,3 +1,2 @@
 export {Legend} from './Legend';
 export {StackedAreas} from './StackedAreas';
-export {VisuallyHiddenRows} from './VisuallyHiddenRows';