+++ src/components/index.ts
@@ -29,3 +29,4 @@ export {
 export {TooltipContent, TooltipContentProps} from './TooltipContent';
 export {BarChartXAxis} from './BarChartXAxis';
 export {SkipLink} from './SkipLink';
+export {VisuallyHiddenRows} from './VisuallyHiddenRows';